Organiza las pantallas de EventDetailFragment y EventOverviewFragment
