+++ app/src/main/java/idnp/grupo_uno/proyecto_final/ui/events/EventDetailFragment.java
@@ -1,0 +1,25 @@
+package idnp.grupo_uno.proyecto_final.ui.events;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import idnp.grupo_uno.proyecto_final.R;
+
+public class EventDetailFragment extends Fragment {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_event_detail, container, false);
+    }
+}
