+++ app/src/main/res/layout/fragment_event_overview.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".EventOverview">
+    tools:context=".ui.events.EventOverviewFragment">
 
     <TextView
         android:layout_width="match_parent"
